maintenance commit
- turned on some warnings for gcc
- updated boost::tribool usages for 1.69
- fixed wrong includes in systemhandler.cpp
- centered controllers are now displayed within range [-64, 63]
- fixed the height of ProgramEditor's table
- removed unnecessary casts within ConfigurationPuller::addConfiguration

+++ src/qcore/manager.cpp
@@ -38,13 +38,13 @@ public:
 
     void addConfiguration(const Configuration& configuration) {
         // set colors
-        for (channel_t c=0 ; c < std::min(channels_t::capacity(), (size_t)configuration.colors.size()) ; ++c)
+        for (channel_t c=0 ; c < std::min(channels_t::capacity(), configuration.colors.size()) ; ++c)
             mManager->channelEditor()->setColor(c, configuration.colors.at(c));
         // add frames
         auto* mainDisplayer = mManager->mainDisplayer();
         if (!configuration.frames.empty())
             setFrame(mainDisplayer, configuration.frames[0], true);
-        for (int i=1 ; i < configuration.frames.size() ; i++)
+        for (size_t i=1 ; i < configuration.frames.size() ; i++)
             addFrame(mainDisplayer, configuration.frames[i], true);
         // add handlers
         for (const auto& handler : configuration.handlers)
